% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_feature.R
\name{getAlt2}
\alias{getAlt2}
\title{Second alternative allele percentage}
\usage{
getAlt2(f)
}
\arguments{
\item{f}{Input raw file}
}
\value{
Percent of the second alternative allele
}
\description{
Second alternative allele percentage
}
