% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_feature.R
\name{getVar}
\alias{getVar}
\title{Calculate zygosity variable}
\usage{
getVar(df, state, hom_mle, het_mle)
}
\arguments{
\item{df}{Input modified file}

\item{state}{Zygosity state}

\item{hom_mle}{MLE in hom model}

\item{het_mle}{MLE in het model}
}
\value{
Zygosity variable
}
\description{
Calculate zygosity variable
}
