% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho_est.R
\name{negll}
\alias{negll}
\title{Negative Log Likelihood}
\usage{
negll(x, size, prob, rho)
}
\arguments{
\item{x}{Depth of alternative allele}

\item{size}{Total depth}

\item{prob}{Theoretical probability for heterozygous is 0.5, for homozygous is 0.999}

\item{rho}{Rho parameter of Beta-Binomial distribution of alternative allele}
}
\description{
Calculates negative log likelihood for beta binomial distribution.
}
