% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vcf_example}
\alias{vcf_example}
\title{VCF example file.}
\format{A list of 4 data frames:}
\source{
Created by Tao Jiang
}
\usage{
vcf_example
}
\description{
An example containing a list of 4 data frames.
}
\keyword{datasets}
